project-semantic-search/requirements.txt
pandas
# sentence-transformers
mysql-connector-python
scikit-learn
